pyproject.toml
[project]
name = "milvus-frontend"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "termtate", email = "meng360842@gmail.com" }
]
dependencies = [
    # common
    "injector>=0.21.0",
    "pandas==1.5.0",
    "pydantic>=2.1.1",
    "openpyxl>=3.1.2",
    "pydantic-settings>=2.0.2",
    "nuitka>=1.7.9",
    # ui
    "qt-material>=2.14",
    "reactivex>=4.0.4",
    "qasync>=0.24.0",
    "pyside6>=6.5.2",
    # network
    "httpx>=0.24.1",
    # ml
    "pytorch-crf>=0.7.2",
    "seqeval==1.2.2",
    "torch==1.12.1",
    "pywin32>=306",
    "pdfminer.six>=20221105",
    "python-docx>=0.8.11",
    "setuptools>=68.0.0",
    "lac>=2.1.2",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.rye.scripts]
pack = [
    "python", "-m",
    "nuitka", "--standalone", "--show-memory", 
    "--show-progress", "--output-dir=out", 
    "--plugin-enable=pyside6", "src/main.py"
]
main = { call = "src.main", env= { PYTHONPATH = "./src/"} }
    
[tool.hatch.metadata]
allow-direct-references = true


[[sources]]
name = "default"
url = "https://mirror.baidu.com/pypi/simple"
